Addresses controller

+++ Vastra.API/Vastra.API/Models/OrderWithoutCartItemsDto.cs
@@ -3,7 +3,7 @@
     public class OrderWithoutCartItemsDto
     {
         public int OrderId { get; set; }
-        public int Value { get; set; }
+        public float Value { get; set; }
         public string PaymentStatus { get; set; }
         public DateTime DateAdded { get; set; }
         public DateTime DateModified { get; set; }
